Moved both filters to Molecules.
