+++ docroot/themes/contrib/civictheme/templates/form/radios--civictheme-basic-filter.html.twig
@@ -12,7 +12,7 @@
  * @ingroup themeable
  */
 #}
-{% embed '@organisms/single-filter/single-filter.twig' with {
+{% embed '@molecules/single-filter/single-filter.twig' with {
   theme: 'light'
 }%}
   {% block filters %}
