+++ docroot/themes/contrib/civictheme/templates/views/views-exposed-form.html.twig
@@ -13,7 +13,7 @@
 {% endif %}
 {# Group filter is printed if there is more than 1 exposed filter in the view #}
 {% if group_filter %}
-  {% include '@organisms/group-filter/group-filter.twig' with {
+  {% include '@molecules/group-filter/group-filter.twig' with {
     filter_title: filter_title,
     desktop_breakpoint: '>=m',
     filters: form,
